requirements.txt
click==7.1.2
flake8==3.9.2
importlib-metadata==4.6.3
Markdown==3.3.4
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyYAML==5.4.1
typer==0.3.2
typing-extensions==3.10.0.0
zipp==3.5.0
